gesture_recognition/requirements.txt
mediapipe==0.10.5
opencv-python==4.8.0.76
numpy==1.23.5
tensorflow==2.12.0
scikit-learn==1.2.2
websockets==12.0
aiohttp==3.9.3
aiosqlite==0.20.0
bcrypt==4.1.3
PyJWT==2.8.0
matplotlib==3.9.2